Move getSwapTransactions to uniV2Utils.js
