+++ src/swap/spookySwap/spookySwap.js
@@ -1,6 +1,5 @@
 // @flow
 
-import { mul, sub } from 'biggystring'
 import {
   type EdgeCorePluginOptions,
   type EdgeSpendInfo,
@@ -11,20 +10,14 @@ import {
   type EdgeSwapResult,
   type EdgeTransaction
 } from 'edge-core-js/types'
-import { type PopulatedTransaction, Contract, ethers } from 'ethers'
+import { ethers } from 'ethers'
 
-import { round } from '../../util/biggystringplus.js'
+import { getInOutTokenAddresses } from '../defi/defiUtils.js'
+import { spookySwapRouter } from '../defi/uni-v2-based/uniV2Contracts.js'
 import {
-  getInOutTokenAddresses,
-  getMetaTokenAddress
-} from '../defi/defiUtils.js'
-import {
-  makeErc20Contract,
-  provider,
-  spookySwapRouter,
-  wrappedFtmToken
-} from '../defi/uni-v2-based/uniV2Contracts.js'
-import { getSwapAmounts } from '../defi/uni-v2-based/uniV2Utils.js'
+  getSwapAmounts,
+  getSwapTransactions
+} from '../defi/uni-v2-based/uniV2Utils.js'
 
 const swapInfo: EdgeSwapInfo = {
   pluginId: 'spookySwap',
@@ -34,143 +27,6 @@ const swapInfo: EdgeSwapInfo = {
 }
 const EXPIRATION_MS = 1000 * 20 * 60
 const SLIPPAGE = '0.05' // 5%
-
-/**
- * Get smart contract transaction(s) necessary to swap based on swap params
- */
-export const getSwapTransactions = async (
-  swapRequest: EdgeSwapRequest,
-  router: Contract,
-  amountToSwap: string,
-  expectedAmountOut: string,
-  toAddress: string,
-  slippage: string,
-  deadline: number
-): Promise<PopulatedTransaction[]> => {
-  const { fromWallet, fromCurrencyCode, toCurrencyCode } = swapRequest
-  const {
-    currencyCode: nativeCurrencyCode,
-    metaTokens
-  } = fromWallet.currencyInfo
-  const fromAddress = (await fromWallet.getReceiveAddress()).publicAddress
-
-  // TODO: Use our new denom implementation to get native amounts
-  const wrappedCurrencyCode = `W${nativeCurrencyCode}`
-  const isFromNativeCurrency = fromCurrencyCode === nativeCurrencyCode
-  const isToNativeCurrency = toCurrencyCode === nativeCurrencyCode
-  const isFromWrappedCurrency = fromCurrencyCode === wrappedCurrencyCode
-  const isToWrappedCurrency = toCurrencyCode === wrappedCurrencyCode
-
-  const fromTokenAddress = getMetaTokenAddress(
-    metaTokens,
-    isFromNativeCurrency ? wrappedCurrencyCode : fromCurrencyCode
-  )
-  const toTokenAddress = getMetaTokenAddress(
-    metaTokens,
-    isToNativeCurrency ? wrappedCurrencyCode : toCurrencyCode
-  )
-
-  // Determine router method name and params
-  if (isFromNativeCurrency && isToNativeCurrency)
-    throw new Error('Invalid swap: Cannot swap to the same native currency')
-  const path = [fromTokenAddress, toTokenAddress]
-
-  const gasPrice = await provider.getGasPrice()
-
-  const addressToApproveTxs = async (
-    tokenAddress: string,
-    contractAddress: string
-  ): PopulatedTransaction | void => {
-    const tokenContract = makeErc20Contract(tokenAddress)
-    const allowence = await tokenContract.allowance(
-      fromAddress,
-      contractAddress
-    )
-    if (allowence.sub(amountToSwap).lt(0)) {
-      return tokenContract.populateTransaction.approve(
-        contractAddress,
-        ethers.constants.MaxUint256,
-        { gasLimit: '60000', gasPrice }
-      )
-    }
-  }
-
-  const txs = await (async (): Promise<
-    Array<Promise<PopulatedTransaction> | void>
-  > => {
-    // Deposit native currency for wrapped token
-    if (isFromNativeCurrency && isToWrappedCurrency) {
-      return [
-        wrappedFtmToken.populateTransaction.deposit({
-          gasLimit: '51000',
-          gasPrice,
-          value: amountToSwap
-        })
-      ]
-    }
-    // Withdraw wrapped token for native currency
-    if (isFromWrappedCurrency && isToNativeCurrency) {
-      return [
-        // Deposit Tx
-        wrappedFtmToken.populateTransaction.withdraw(amountToSwap, {
-          gasLimit: '51000',
-          gasPrice
-        })
-      ]
-    }
-    // Swap native currency for token
-
-    const slippageMultiplier = sub('1', slippage)
-    if (isFromNativeCurrency && !isToNativeCurrency) {
-      return [
-        // Swap Tx
-        router.populateTransaction.swapExactETHForTokens(
-          round(mul(expectedAmountOut, slippageMultiplier)),
-          path,
-          toAddress,
-          deadline,
-          { gasLimit: '250000', gasPrice, value: amountToSwap }
-        )
-      ]
-    }
-    // Swap token for native currency
-    if (!isFromNativeCurrency && isToNativeCurrency) {
-      return [
-        // Approve TX
-        await addressToApproveTxs(path[0], router.address),
-        // Swap Tx
-        router.populateTransaction.swapExactTokensForETH(
-          amountToSwap,
-          round(mul(expectedAmountOut, slippageMultiplier)),
-          path,
-          toAddress,
-          deadline,
-          { gasLimit: '250000', gasPrice }
-        )
-      ]
-    }
-    // Swap token for token
-    if (!isFromNativeCurrency && !isToNativeCurrency) {
-      return [
-        // Approve TX
-        await addressToApproveTxs(path[0], router.address),
-        // Swap Tx
-        router.populateTransaction.swapExactTokensForTokens(
-          amountToSwap,
-          round(mul(expectedAmountOut, slippageMultiplier)),
-          path,
-          toAddress,
-          deadline,
-          { gasLimit: '600000', gasPrice }
-        )
-      ]
-    }
-
-    throw new Error('Unhandled swap type')
-  })()
-
-  return await Promise.all(txs.filter(tx => tx != null))
-}
 
 export function makeSpookySwapPlugin(
   opts: EdgeCorePluginOptions
